Attempt to calm the testing gods...
